separate runCommand, runCreate, runEdit, runJest, into ./run
