+++ src/cli/createCommand.ts
@@ -1,62 +1,6 @@
 import program, { Command } from 'commander';
-import { loadConfig } from '../config';
 import { parseUrl } from './parseUrl';
-import { runCommand } from './runCommand';
-import { runJest } from './runJest';
-import { saveTemplate } from './saveTemplate';
-
-export type CreateOptions = {
-  device?: string;
-  isScript?: boolean;
-  name: string;
-  statePath?: string;
-  url: string;
-};
-
-export const runCreate = async (options: CreateOptions): Promise<void> => {
-  const config = loadConfig();
-
-  const codePath = await saveTemplate({
-    device: options.device,
-    isScript: options.isScript,
-    name: options.name,
-    rootDir: config.rootDir,
-    statePath: options.statePath,
-    templateFn: config.createTemplate,
-    url: options.url,
-    useTypeScript: config.useTypeScript,
-  });
-  if (!codePath) {
-    // the user decided to not overwrite
-    return;
-  }
-
-  const env: NodeJS.ProcessEnv = {
-    QAW_CREATE: 'true',
-    QAW_HEADLESS: 'false',
-  };
-
-  if (options.isScript) {
-    runCommand(
-      `${config.useTypeScript ? 'ts-node -D 6133' : 'node'} ${codePath}`,
-      {
-        ...env,
-        QAW_BROWSER: 'chromium',
-      },
-    );
-    return;
-  }
-
-  runJest({
-    browsers: ['chromium'],
-    config: config.config,
-    env,
-    repl: true,
-    rootDir: config.rootDir,
-    testPath: codePath,
-    testTimeout: config.testTimeout,
-  });
-};
+import { runCreate } from '../run/runCreate';
 
 export const buildCreateCommand = (): program.Command => {
   const command = new Command('create')
