+++ test/cli/omitArgs.test.ts
@@ -1,4 +1,4 @@
-import { omitArgs } from '../src/cli/omitArgs';
+import { omitArgs } from '../../src/cli/omitArgs';
 
 it('omits args', () => {
   const args = omitArgs(['--omit=true', '--keep=true'], ['--omit']);
